requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py.git@faffc8eeb2ca06cae60b726fa873e474a51ac39b
psycopg2-binary